+++ packages/orq-rc/src/models/components/conversationwithmessagesresponse.ts
@@ -155,7 +155,7 @@ export const ConversationWithMessagesResponse$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kevxwajm645xq1zrem5zwcv2"),
+  _id: z.string().default("conv_01kexrj7rva680qs5hyhz2pam3"),
   kind: ConversationWithMessagesResponseKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),
@@ -189,7 +189,7 @@ export const ConversationWithMessagesResponse$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ConversationWithMessagesResponse
 > = z.object({
-  id: z.string().default("conv_01kevxwajm645xq1zrem5zwcv2"),
+  id: z.string().default("conv_01kexrj7rva680qs5hyhz2pam3"),
   kind: ConversationWithMessagesResponseKind$outboundSchema,
   displayName: z.string(),
   createdAt: z.number(),