+++ packages/orq-rc/src/models/operations/generateconversationname.ts
@@ -251,7 +251,7 @@ export const GenerateConversationNameResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kevxwajbf0cyqfjjvkg3ec8k"),
+  _id: z.string().default("conv_01kexrj7rsncjwv8htv64hpv94"),
   kind: GenerateConversationNameKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),
@@ -283,7 +283,7 @@ export const GenerateConversationNameResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   GenerateConversationNameResponseBody
 > = z.object({
-  id: z.string().default("conv_01kevxwajbf0cyqfjjvkg3ec8k"),
+  id: z.string().default("conv_01kexrj7rsncjwv8htv64hpv94"),
   kind: GenerateConversationNameKind$outboundSchema,
   displayName: z.string(),
   createdAt: z.number(),