+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -4633,7 +4633,7 @@ export const AgentToolInputRunTools$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01KEVXWAG334R44SJ09RG598T1"),
+  id: z.string().default("01KEXRJ7NDSYG7V3YEZKK55TE9"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$inboundSchema),
@@ -4652,7 +4652,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01KEVXWAG334R44SJ09RG598T1"),
+  id: z.string().default("01KEXRJ7NDSYG7V3YEZKK55TE9"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),